+++ Game/InventoryMenu.py
@@ -3,32 +3,23 @@ from curses import panel
 from .Menu import Menu
 
 class InventoryMenu(Menu):
-
-    
  #   def __init__(self, title, items, stdscreen):
-
-
     def display(self):
             self.panel.top()
             self.panel.show()
             self.window.clear()
-
             offsetY = 2
             offsetX = 2
-
             while True:
                 self.window.refresh()
                 curses.doupdate()
-
                 #self.window.addstr( 0, 0, self.title )
-
                 self.window.border('|','|','-','-','+','+','+','+')
                 offsetY = 2
                 offsetX = 2
                 self.window.addstr( offsetY, offsetX, self.title )
                 offsetX += 2
                 offsetY += 2
-
 
                 for index, item in enumerate(self.items):
                     if index == self.position: